Stubbed in method to actually move the file. Added a bunch of assertions there, but not actually doing the work yet.
Added a save panel for the case of clicking rename without selecting a destination directory. It's possible that we'll need to use the save panel when we have a destination directory to keep the sandbox happy, but let's try to skip that for now.

Added a bunch of notes to self on details to take care of, both UX and handling of file extensions.

Added a helper method for getting the URL of the current source list selection.