Now correctly updating the preview view after renaming a file with multiple files in the source pane.
We don't get the selection-changed notification for the source view when we set it to the same row number that was selected. That happens when we rename a file with multiple files in the source pane.

Now when we remove a file from the source list, besides updating the selection, we also call helper methods to update the preview view and the enabled state.